Add validation

+++ UsersCrud.Api/Controllers/UserController.cs
@@ -11,13 +11,16 @@ namespace UsersCrud.Api.Controllers
     public class UserController : ControllerBase
     {
         private readonly IUserService _userService;
+        private readonly IRoleService _roleService;
         private readonly IMapper _mapper;
 
         public UserController(
             IUserService userService,
+            IRoleService roleService,
             IMapper mapper)
         {
             _userService = userService;
+            _roleService = roleService;
             _mapper = mapper;
         }
 
@@ -45,7 +48,7 @@ namespace UsersCrud.Api.Controllers
         }
 
         [HttpPost]
-        public async Task<UserDto> AddUserAsync(CreateUserDto userDto)
+        public async Task<UserDto> AddUserAsync(UpdateUserDto userDto)
         {
             var user = _mapper.Map<UserModel>(userDto);
             var addedUser = await _userService.AddUserAsync(user);
@@ -57,25 +60,55 @@ namespace UsersCrud.Api.Controllers
         [HttpPut("{userId}/role/{roleId}")]
         public async Task<IActionResult> AddRoleToUserAsync(Guid userId, Guid roleId)
         {
-            var user = await _userService.AddRoleToUserAsync(userId, roleId);
-            var result = _mapper.Map<UserDto>(user);
+            var user = await _userService.GetUserByIdAsync(userId);
+            if (user == null)
+            {
+                return NotFound(userId);
+            }
+
+            var checkRole = user.Roles.Any(x => x.Id == roleId);
+            if (checkRole)
+            {
+                return BadRequest("Role already exists");
+            }
+
+            var role = await _roleService.GetRoleByIdAsync(roleId);
+            if (role == null)
+            {
+                return NotFound(roleId);
+            }
+
+            var userModel = await _userService.AddRoleToUserAsync(userId, roleId);
+            var result = _mapper.Map<UserDto>(userModel);
 
             return Ok(result);
         }
 
-        [HttpPut]
-        public async Task<UserDto> UpdateUserAsync(UpdateUserDto userDto)
+        [HttpPut("{id}")]
+        public async Task<IActionResult> UpdateUserAsync(Guid id, UpdateUserDto userDto)
         {
-            var user = _mapper.Map<UserModel>(userDto);
-            var updatedUser = await _userService.UpdateUserAsync(user);
+            var user = await _userService.GetUserByIdAsync(id);
+            if (user == null)
+            {
+                return NotFound();
+            }
+
+            var userModel = _mapper.Map<UserModel>(userDto);
+            var updatedUser = await _userService.UpdateUserAsync(userModel);
             var result = _mapper.Map<UserDto>(updatedUser);
 
-            return result;
+            return Ok(result);
         }
 
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteUserAsync(Guid id)
         {
+            var user = await _userService.GetUserByIdAsync(id);
+            if (user == null)
+            {
+                return NotFound();
+            }
+
             await _userService.RemoveUserAsync(id);
 
             return Ok();
